+++ src/main/java/org/sopt/repository/PostRepository.java
@@ -1,38 +1,17 @@
 package org.sopt.repository;
 
 import org.sopt.domain.Post;
-import org.sopt.exception.PostNotFoundException;
-import org.springframework.stereotype.Repository;
 
-import java.util.ArrayList;
 import java.util.List;
 
-@Repository
-public class PostRepository {
-    private static final List<Post> postList = new ArrayList<>();
-    private Long nextId = 1L;
+public interface PostRepository {
+    Post save(Post post);
 
-    public Post save(Post post) {
-        postList.add(post);
-        return post;
-    }
+    Long generateId();
 
-    public Long generateId() {
-        return nextId++;
-    }
+    Post findById(Long id);
 
-    public Post findById(Long id) {  // null 대신 Optional
-        return postList.stream()
-                .filter(p -> p.getId().equals(id))
-                .findFirst()
-                .orElseThrow(PostNotFoundException::new);
-    }
+    List<Post> getAllPosts();
 
-    public List<Post> getAllPosts() {
-        return postList;
-    }
-
-    public void remove(Post post) {
-        postList.remove(post);
-    }
+    void remove(Post post);
 }
